// api/resources/assets/client/src/app/models/saveCampaign.ts
export class SaveCampaign {
  id: number;
  unique_script_id: string;
  created_by: number;
  campaign_name: string;
  domain_name: string;
  styles: string;
  delay: number;
  delay_before_start: number;
  loop: any;
  branding: number;
  branding_id: number;
  created_at: any;
  updated_at: any;
  deleted_at: any;
  is_active: string;
  exit_pop_up: any;
  exit_pop_up_id: number;
}